Fixed bug where leaving building would crash program
This problem was caused when the stack destructed the unique_ptr, the
unique_ptr then tried to delete the pointer object, but the pointer
object was owned by another unique_ptr in the building array. Solution
was to release ownership of the pointer before popping the stack.
Shared_ptr would technically be safer, but the stack is of type
unique_ptr with functionality implemented already with the game states.